Basic sankey without labels
