+++ sankey-trash/index.html
@@ -1,17 +1,17 @@
 <!DOCTYPE html>
 <html>
 <head>
-    <title>Sankey diagram — Visual Vocabulary</title>
+    <title>Wrapper-starter Visual Vocabulary</title>
 
-    <script src="//unpkg.com/@financial-times/d3-bootloader" async></script>
+    <script src="//unpkg.com/@financial-times/d3-bootloader@^1.6.0" async></script>
     <link rel="stylesheet" href="https://ig.ft.com/visual-vocabulary/styles.css"/>
     <link rel="stylesheet" href="//www.ft.com/__origami/service/build/v2/bundles/css?modules=o-fonts@^2.2.0" />
 </head>
 <body>
 
-    <h1>Sankey diagram</h1>
-    <p>The standard way to show a changing time series. If data are irregular, consider markers to represent data points</p>
-    <p><b>Some useful settings in the index.js</b>
+    <h1>Sankey</h1>
+    <p>A useful way of showing non-cyclical transfer from one entity to another</p>
+    <!-- <p><b>Some useful settings in the index.js</b>
         <ul>
             <li><b>yMin</b> and <b>yMax</b> – use these to set the lower and upper values of the y-axis</li>
             <li><b>numTicksy</b> – set the number of scale lines on the y-axis</li>
@@ -19,14 +19,13 @@
             <li><b>highlightNames</b> – put the column names of the data you want to highlight, all other series will be muted</li>
             <li><b>interval</b> – options available are "century","jubilee" (50 years), decade", "lustrum" (5 years), "years","months","days"</li>
         </ul>
-    </p>
+    </p> -->
      <p><b>Column headings in the csv</b>
         <ul>
             <li><b>annotate</b> – use this column to create annotations for a particular point in time</li>
             <li><b>highlight</b> – use this column to define highlight bands in the background of your chart, add 'begin' where you want the band to start and 'end' where you want it to finish. You can have as many highlights as you want.</li>
         </ul>
     </p>
-    <p>Please make sure than the <b>blue line</b> is on top as this is the primary line colour</p>
 
     <hr>
